+++ CHANGELOG.md
@@ -23,6 +23,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.1.0] - 2018-01-24
+### Changed
+- ASG: Variable ingress rules for ASG security group
+- ASG with ELB: Variable ingress rules for ELB security group
+
 ## [0.0.9] - 2018-01-23
 ### Removed
 - ASG with ELB: Removed unused variables
@@ -59,7 +64,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## 0.0.1 - 2018-01-17
 Starting tag
 
-[Unreleased]: https://github.com/albumprinter/payments-tf-modules/compare/v0.0.9...HEAD
+[Unreleased]: https://github.com/albumprinter/payments-tf-modules/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/albumprinter/payments-tf-modules/compare/v0.0.9...v0.1.0
 [0.0.9]: https://github.com/albumprinter/payments-tf-modules/compare/v0.0.8...v0.0.9
 [0.0.8]: https://github.com/albumprinter/payments-tf-modules/compare/v0.0.7...v0.0.8
 [0.0.7]: https://github.com/albumprinter/payments-tf-modules/compare/v0.0.6...v0.0.7